made mega_check a function of the proof itself
