+++ chain-vote/src/gang/ristretto255.rs
@@ -12,147 +12,9 @@ use rand_core::{CryptoRng, RngCore};
 use std::hash::{Hash, Hasher};
 use std::ops::{Add, Mul, Sub};
 
-use crate::encrypted::PTP;
-use crate::private_voting::unit_vector_zkp::ResponseRandomness;
-use crate::unit_vector::binrep;
-use crate::Ciphertext;
+use curve25519_dalek_ng::traits::VartimeMultiscalarMul;
 use std::array::TryFromSliceError;
 use std::convert::TryInto;
-use std::iter;
-use crate::encryption::PublicKey;
-use curve25519_dalek_ng::traits::VartimeMultiscalarMul;
-use crate::private_voting::Announcement;
-
-pub(crate) fn mega_check(
-    ciphertexts: &PTP<Ciphertext>,
-    public_key: &PublicKey,
-    challenge_x: &Scalar,
-    committed_rand: &[ResponseRandomness],
-    challenge_y: &Scalar,
-    encrypted_coeff: &[Ciphertext],
-    response: &Scalar,
-    ibas: &[Announcement],
-) -> bool {
-
-    let bits = ciphertexts.bits();
-    let length = ciphertexts.len();
-    let cx_pow = challenge_x.power(bits);
-
-    let powers_cx = exp_iter(*challenge_x);
-    let powers_cy = exp_iter(*challenge_y);
-
-    let powers_z_iterator = powers_z_encs_iter(committed_rand, challenge_x, &(bits as u32));
-
-    let zero = public_key.encrypt_with_r(&Scalar::zero(), &response);
-
-    // check commitments are 0 / 1
-    for (iba, zwv) in self.ibas.iter().zip(self.zwvs.iter()) {
-        let com1 = ck.commit(&zwv.z, &zwv.w);
-        let lhs = &iba.i * &cx + &iba.b;
-        if lhs != com1 {
-            return false;
-        }
-
-        let com2 = ck.commit(&Scalar::zero(), &zwv.v);
-        let lhs = &iba.i * (&cx - &zwv.z) + &iba.a;
-        if lhs != com2 {
-            return false;
-        }
-    }
-
-    let mega_check = Point::optional_multiscalar_mul(
-        powers_cy.clone().take(length).map(|s| s.0 * cx_pow.0)
-            .chain(powers_cy.clone().take(length).map(|s| s.0 * cx_pow.0))
-            .chain(powers_cy.take(length).map(|s| s.0))
-            .chain(powers_cx.clone().take(bits).map(|s| s.0))
-            .chain(powers_cx.take(bits).map(|s| s.0))
-            .chain(iter::once(- IScalar::one()))
-            .chain(iter::once(- IScalar::one())),
-        ciphertexts.iter().map(|ctxt| Some(ctxt.e2.0))
-            .chain(ciphertexts.iter().map(|ctxt| Some(ctxt.e1.0)))
-            .chain(powers_z_iterator.take(length).map(|p| Some(p)))
-            .chain(encrypted_coeff.iter().map(|ctxt| Some(ctxt.e1.0)))
-            .chain(encrypted_coeff.iter().map(|ctxt| Some(ctxt.e2.0)))
-            .chain(iter::once(Some(zero.e1.0)))
-            .chain(iter::once(Some(zero.e2.0))),
-    ).expect("This went terribly wrong");
-
-    mega_check == Point::identity()
-}
-
-/// Computes the product of the power of `z` given an `index` and a `bit_size`
-fn powers_z_encs(z: &[ResponseRandomness], challenge_x: Scalar, index: usize, bit_size: u32) -> Scalar{
-    println!("index: {}", index);
-    let idx = binrep(index, bit_size as u32);
-
-    let multz =
-        z
-            .iter()
-            .enumerate()
-            .fold(Scalar::one(), |acc, (j, zwv)| {
-                let m = if idx[j] { zwv.z.clone() } else { challenge_x - &zwv.z };
-                &acc * m
-            });
-    multz
-}
-
-/// Provides an iterator over the powers of a `Scalar`.
-///
-/// This struct is created by the `exp_iter` function.
-pub struct ZPowExp {
-    index: usize,
-    bit_size: u32,
-    z: Vec<ResponseRandomness>,
-    challenge_x: Scalar,
-}
-
-impl Iterator for ZPowExp {
-    type Item = Point;
-
-    fn next(&mut self) -> Option<Point> {
-        let z_pow = powers_z_encs(&self.z, self.challenge_x, self.index, self.bit_size);
-        self.index += 1;
-        Some(z_pow.negate().0 * RISTRETTO_BASEPOINT_POINT)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (usize::MAX, None)
-    }
-}
-
-/// Return an iterator of the powers of `ZPowExp`.
-fn powers_z_encs_iter(z: &[ResponseRandomness], challenge_x: &Scalar, bit_size: &u32) -> ZPowExp {
-    ZPowExp { index: 0, bit_size: bit_size.clone(),z:  z.to_vec(), challenge_x: challenge_x.clone()}
-}
-
-/// Provides an iterator over the powers of a `Scalar`.
-///
-/// This struct is created by the `exp_iter` function.
-#[derive(Clone)]
-pub struct ScalarExp {
-    x: Scalar,
-    next_exp_x: Scalar,
-}
-
-impl Iterator for ScalarExp {
-    type Item = Scalar;
-
-    fn next(&mut self) -> Option<Scalar> {
-        let exp_x = self.next_exp_x;
-        self.next_exp_x = self.next_exp_x * self.x;
-        Some(exp_x)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (usize::MAX, None)
-    }
-}
-
-/// Return an iterator of the powers of `x`.
-fn exp_iter(x: Scalar) -> ScalarExp {
-    let next_exp_x = Scalar::one();
-    ScalarExp { x, next_exp_x }
-}
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct Scalar(IScalar);
@@ -218,6 +80,16 @@ impl GroupElement {
             sum = sum + v;
         }
         sum
+    }
+    pub fn multiscalar_multiplication<I, J>(scalars: I, points: J) -> Self
+    where
+        I: IntoIterator<Item = Scalar>,
+        J: IntoIterator<Item = GroupElement>,
+    {
+        GroupElement(Point::vartime_multiscalar_mul(
+            scalars.into_iter().map(|s| s.0),
+            points.into_iter().map(|p| p.0),
+        ))
     }
 }
 
@@ -294,6 +166,38 @@ impl Scalar {
             sum = sum + v;
         }
         Some(sum)
+    }
+
+    /// Return an iterator of the powers of `x`.
+    pub fn exp_iter(&self) -> ScalarExp {
+        let next_exp_x = Scalar::one();
+        ScalarExp {
+            x: self.clone(),
+            next_exp_x,
+        }
+    }
+}
+
+/// Provides an iterator over the powers of a `Scalar`.
+///
+/// This struct is created by the `exp_iter` function.
+#[derive(Clone, Copy)]
+pub struct ScalarExp {
+    x: Scalar,
+    next_exp_x: Scalar,
+}
+
+impl Iterator for ScalarExp {
+    type Item = Scalar;
+
+    fn next(&mut self) -> Option<Scalar> {
+        let exp_x = self.next_exp_x;
+        self.next_exp_x = self.next_exp_x * self.x;
+        Some(exp_x)
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (usize::MAX, None)
     }
 }
