+++ chain-vote/src/private_voting/mod.rs
@@ -10,8 +10,9 @@
 //! the encrypted tuple.
 
 pub(crate) mod challenge_context;
+pub mod messages;
 pub mod unit_vector_zkp;
 
 pub(crate) use self::{
-    challenge_context::ChallengeContext, unit_vector_zkp::Announcement, unit_vector_zkp::Proof,
+    challenge_context::ChallengeContext, messages::Announcement, unit_vector_zkp::Proof,
 };
